Spell Creation
Spells are now cast on the press (Q, E, or Space).

+ Spells are now cast.
+ SpellBook prefab is part of Player now. SpellBook holds the transforms
for all the spells.

! Spells do not rotate with player, they always fire down.

+++ To Kill a Dragon/Assets/Scripts/Spells/Spell.cs
@@ -4,8 +4,8 @@ using System.Collections;
 public class Spell : MonoBehaviour {
 
 	public int SpellNumber;
-	public int SpellCost;
-	public int CastDuration;
+	public float SpellCost;
+	public float CastDuration;
 	public int CastDelay;
 
 	// Use this for initialization
